requirements.txt
pyTelegramBotAPI
aiohttp
SQLAlchemy
sqlalchemy[asyncio]
pillow
pyzbar
aiosqlite